+++ NEWS.rst
@@ -10,7 +10,10 @@ Version 0.8.9 - dev
           entities in all layouts and blocks
    - NEW: Drawing.query(), entity query over all layouts and blocks
    - NEW: Drawing.groupby(), groups DXF entities of all layouts and blocks by an DXF attribute or a key function
-   - CHANGE: Drawing.block.delete_block(name, safe=True), new parameter save, check if block is still referenced (raises DXFValueError)
+   - NEW: Layout.set_redraw_order() and Layout.get_redraw_order(), to change redraw order of entities in model space and
+     paper space layouts
+   - CHANGE: Drawing.block.delete_block(name, safe=True), new parameter save, check if block is still referenced
+     (raises DXFValueError)
    - CHANGE: Drawing.block.delete_all_block(safe=True), new parameter save, ignores blocks still referenced if safe is True
    - NEW: BlockLayout.is_layout_block, True if block is a model space or paper space block definition
    - `Basic` read support for almost all missing DXF entities/objects: